+++ CJBCheatsMenu/docs/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Added 1,000,000g option to the add-money cheat.
 * Added comma separators to add-money and add-casino-coin cheats.
+* Fixed some cases where the weather tab showed the wrong weather for tomorrow.
 * Improved translations. Thanks to CaranudLapin (updated French), insaneArian (updated Polish), and therealmate (updated Hungarian)!
 
 ## 1.38.0